requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
Flask-JWT-Extended>=4.4
Flask-Marshmallow
marshmallow
psycopg2-binary
python-dotenv
pytest
Flask-Testing
flasgger
marshmallow-sqlalchemy
